+++ pocketutils/logging/log_format.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import logging
+import textwrap
 from copy import deepcopy
 
 
@@ -88,25 +89,43 @@ class PrettyRecordFactory:
         log_factory = KaleRecordFactory(7, 13, 5).modifying(logger)
 
     For example:
-        [20191228:14:20:06] kale⟩    datasets      :77    INFO    | Downloading QC-DR...
-        [20191228:14:20:08] numexp…⟩ utils         :141   INFO    | NumExpr defaulting to 8 threads.
-        [20191228:14:21:01] kale⟩    __init__      :185   NOTICE  | Downloaded QC-DR with 8 runs, 85 names, and 768 wells.
-        [20191229:14:26:04] kale⟩    __init__      :202   INFO    | Registered new type RandomForestClassifier:n_jobs=4,n_estimators=8000
-        [20191229:14:26:36] kale⟩    __init__      :185   NOTICE  | Using: treatment selector ⟨keep_all⟩, control selector ⟨keep_all⟩, subsampler ⟨keep_8_balanced⟩, and should_proceed ⟨if_at_least_0_req2_fail⟩
-        [20191229:14:26:36] kale⟩    multi_traine… :193   DEBUG   | Logging every 50 iterations.
-        [20191229:14:26:36] kale⟩    timing        :83    INFO    | Started processing 504 items at 2019-12-28 14:26:36.
-        [20191229:14:26:39] kale⟩    timing        :91    INFO    | Processed 1/504 in 2.8s. Estimated 23.57min left.
-        [20191229:14:26:41] kale⟩    classifiers   :211   INFO    | Training on 2 labels and 101997 features using 16 examples, 8 runs, and 2000 estimators on 4 core(s).
-        [20191229:14:26:46] kale⟩    classifiers   :230   INFO    | Finished training. Took 5s. oob_score=0.188
-        [20191229:14:27:07] kale⟩    __init__      :185   NOTICE  | Ignoring future classifier output...
-        [20191229:14:39:13] kale⟩    timing        :91    INFO    | Processed 201/504 in 12.61min. Estimated 19.01min left.
-        [20191229:14:51:12] kale⟩    timing        :91    INFO    | Processed 401/504 in 24.6min. Estimated 6.32min left.
-        [20191229:14:57:10] kale⟩    timing        :94    INFO    | Processed 504/504 in 30.58min. Done at 2019-12-28 14:57:10.
+    •  ⟩20210302:17:33:14 :sauronlab  :environment  :150
+       |Set global log level to INFO and sauronlab to MINOR.
+    •  ⟩20210302:17:33:14 :sauronlab  :environment  :141
+       |Read Csauronlab.config .
+    •  ⟩20210302:17:33:14 :sauronlab  :environment  :142
+       |Set 8 sauronlab config items. Run 'print(sauronlab_env.info())' for details.
+    ⚠  ⟩20210302:17:33:22 :sauronlab  :startup      :26
+       |Could not load sauronlab package info. Is it installed?
+    •  ⟩20210302:17:33:26 :sauronlab  :_kvrc_utils  :989
+       |Loaded 49 matplotlib RC settings from sauronlab.mplstyle
+    •  ⟩20210302:17:33:26 :sauronlab  :_kvrc_utils  :1035
+       |Loaded 37 Sauronlab viz settings from sauronlab_viz.properties
+    •  ⟩20210302:17:33:26 :sauronlab  :_kvrc_utils  :1037
+       |Set 10 reference widths and heights. Pad is (0.25, 0.25). Gutter is 0.25.
+    ★  ⟩20210302:17:33:26 :sauronlab  :fancy_logger :203
+       |Sauronlab version ??. Started in 12s.
     """
 
-    def __init__(self, max_name: int, max_module: int, max_line: int):
+    def __init__(
+        self, max_name: int, max_module: int, max_line: int, width: int = 100, symbols: bool = False
+    ):
         self.max_name, self.max_module, self.max_line = max_name, max_module, max_line
+        self.width = width
         self.old_factory = deepcopy(logging.getLogRecordFactory())
+        if symbols:
+            self.flags = dict(
+                TRACE=".",
+                DEBUG="-",
+                MINOR="·",
+                INFO="•",
+                CAUTION="☡",
+                WARNING="⚠",
+                NOTICE="★",
+                ERROR="⛔",
+            )
+        else:
+            self.flags = {}
 
     def factory(self, *args, **kwargs):
         def ltrunc(s, n, before="", after=""):
@@ -115,14 +134,31 @@ class PrettyRecordFactory:
             )
 
         record = self.old_factory(*args, **kwargs)
+        if record.name == "py.warnings":
+            record.name = "warn"
+        if record.module.startswith("<ipython-input"):
+            record.module = "ipython"
+        record.symbol = self.flags.get(record.levelname, record.levelname)
         record.abbrev = (
-            ltrunc(record.name, self.max_name, after="⟩")
-            + " "
+            ":"
+            + ltrunc(record.name, self.max_name)
+            + " :"
             + ltrunc(record.module, self.max_module)
-            + " "
-            + ":"
+            + " :"
             + str(record.lineno).ljust(self.max_line)
         )
+        msg = str(record.getMessage())
+        # TODO: 68
+        if len(msg) < self.width - 64:
+            record.wrapped = " " + msg
+        else:
+            # 18 from the timestamp
+            hanging = " " * 3 + "|"
+            record.wrapped = "\n" + "\n".join(
+                textwrap.wrap(
+                    msg, width=self.width, initial_indent=hanging, subsequent_indent=hanging
+                )
+            )
         return record
 
     @property
@@ -131,7 +167,7 @@ class PrettyRecordFactory:
 
     @property
     def format(self) -> str:
-        return "[%(asctime)s] %(abbrev)s %(levelname)-8s| %(message)s"
+        return "%(symbol)s  ⟩%(asctime)s %(abbrev)s%(wrapped)s"
 
     @property
     def time_format(self) -> str: